requirements.txt
gdown==4.5.3
matplotlib==3.6.0
numpy==1.23.3
pandas==1.5.0
Pillow==9.4.0
torch==1.13.1
tqdm==4.64.1
transformers==4.22.1